rename project: mds -> pytype
